+++ .settings/org.eclipse.wst.common.component
@@ -1,8 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
+        
     <wb-module deploy-name="banchan">
+                
         <wb-resource deploy-path="/" source-path="/src/main/webapp" tag="defaultRootSource"/>
+                
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/java"/>
-        <property name="context-root" value="banchan"/>
+                
         <property name="java-output-path" value="/banchan/build/classes"/>
+        <property name="context-root" value="/"/>
+            
     </wb-module>
+    
 </project-modules>
