+++ src/main/java/com/blacksmith/banchan/admin/action/AdminLoginAction.java
@@ -28,7 +28,8 @@ public class AdminLoginAction implements Action {
 
         
         String adminId = request.getParameter("adminId");
-        String adminPassword = passwordHashing.hashPassword(request.getParameter("adminPassword"));
+//        String adminPassword = passwordHashing.hashPassword(request.getParameter("adminPassword"));
+        String adminPassword = request.getParameter("adminPassword");
         
         int result = adminDAO.adminCheck(adminId, adminPassword);
