+++ src/main/java/com/blacksmith/banchan/login/action/LogoutAction.java
@@ -16,12 +16,13 @@ public class LogoutAction implements Action {
 	public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		HttpSession session = request.getSession(false);
 		String jwtToken = request.getHeader("Authorization");
+        System.out.println(session);
         System.out.println(jwtToken);
         if (session != null) {
             // 세션에서 토큰을 삭제합니다.
             session.removeAttribute("Authorization");
             session.invalidate(); // 옵션: 세션 자체를 무효화합니다.
-            System.out.println(jwtToken);
+            
             // 클라이언트에게 로그아웃 성공 메시지를 응답합니다.
             response.getWriter().write("Logout successful");
             response.setStatus(HttpServletResponse.SC_OK);
