+++ src/main/java/com/blacksmith/banchan/util/CharactorEncodingFilter.java
@@ -15,7 +15,9 @@ public class CharactorEncodingFilter implements Filter {
 			, ServletResponse response
 			, FilterChain chain)
 			throws IOException, ServletException {
-		System.out.println("filter1");	
+		
+		
+		System.out.println("hfjshfgsajfsajkfhsfjadddddddddddddddddddddddddddd");
 		request.setCharacterEncoding("UTF-8");
 		chain.doFilter(request, response);
