+++ src/main/java/com/blacksmith/banchan/util/TestToken.java
@@ -1,0 +1,13 @@
+package com.blacksmith.banchan.util;
+
+public class TestToken {
+
+	public static void main(String[] args) {
+//		
+//		String ok = TokenGenerator.generateJwtToken();
+//		
+//		System.out.println(ok);
+
+	}
+
+}
