--- src/main/webapp/NewFile.html
@@ -1,44 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <script src="resources/js/authorization.js"></script>
-    <title>Send Cart Items</title>
-</head>
-<body>
-    <button onclick="buySelectedItems()">한글 안 깨지나?</button>
-    <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-    <script>
-    function buySelectedItems() {
-        var cartItems = JSON.parse(sessionStorage.getItem('cartItems'));
-        // 서버로 데이터 전송하기 (예: AJAX)
-        $.ajax({
-            type: "POST",
-            url: "banchan?command=test",
-            data: JSON.stringify(cartItems),
-            contentType: "application/json",
-            success: function (response) {
-                console.log("Data sent successfully:", response);
-                // 성공적으로 데이터를 서버로 보낸 후에 수행할 작업
-            },
-            error: function (xhr, status, error) {
-                console.error("Error sending data:", error);
-                // 데이터 전송에 실패한 경우 수행할 작업
-            }
-        });
-    }
-    </script>
-</body>
-	<script>
-	
-    // 세션 스토리지에서 토큰을 가져옴
-    var adminToken = sessionStorage.getItem("adminToken");
-    
-    // 토큰이 없으면 로그인 페이지로 리다이렉트
-    if (!adminToken) {
-        window.location.href = "banchan?command=admin_login_form";
-   
-}
-	</script>
-</html>
