+++ src/main/webapp/WEB-INF/web.xml
@@ -14,17 +14,6 @@
 		<url-pattern>/admin/*</url-pattern>
 	</filter-mapping>
 	
-	<filter>
-		<filter-name>UploadDirectoryFilter</filter-name>
-		<filter-class>com.blacksmith.banchan.util.UploadDirectoryFilter</filter-class>
-	</filter>
-
-	<filter-mapping>
-		<filter-name>UploadDirectoryFilter</filter-name>
-		<url-pattern>/views/app.jsp</url-pattern>
-	</filter-mapping>
-	
-	
 	
 		<filter>
 	    <filter-name>CharactorEncodingFilter</filter-name>
@@ -42,7 +31,10 @@
 	<welcome-file-list>
 		<welcome-file>views/app.jsp</welcome-file>
 	</welcome-file-list>
-
+	
+	<listener>
+        <listener-class>com.blacksmith.banchan.util.InitializationListener</listener-class>
+    </listener>
 	
 
 </web-app>
