+++ src/main/webapp/resources/css/main.css
@@ -205,6 +205,7 @@ header>#bottomHeader>nav>ul>li>a {
     font-weight: 600;
     padding: 12px 0;
     font-family: inherit;
+    line-height: normal;
 }
 
 header>#bottomHeader>p {
@@ -225,42 +226,42 @@ header>#bottomHeader>p>span>a {
 }
 
 
+/* navi */
+.jingaTitle {
+    background-color: white;
+    color: black;
+    padding: 5px;
+    display: block;
+    text-decoration: none;
+}
+
+.jingaTitle:hover {
+    background-color: red;
+    color: white;
+}
 
 
 /* slider */
-#slider {
-    display: flex;
+
+.fade-in-out-slider {
+    position: relative;
     width: 100%;
-    height: 414px;
-    background-color: #FAF7F5;
-    align-items: center;
-    justify-content: center;
+    height: 300px; /* You can adjust the height as needed */
+    overflow: hidden;
 }
-
-#slider>img {
+.fade-in-out-slider img {
     position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    opacity: 0;
+    transition: opacity 2s ease-in-out;
 }
-
-#mobileAward {
+.fade-in-out-slider img.active {
     opacity: 1;
 }
-
-#delivery {
-    opacity: 1;
-}
-
-#slider>ul>li {
-    float: left;
-}
-
-#slider>ul>li> {
-    color: #ffffff;
-    position: absolute;
-
-}
-
-
-
 
 
 /*footer*/
