+++ src/main/webapp/views/admin/admin.jsp
@@ -25,5 +25,15 @@
         <a href="admin?command=review_list" class="btn btn-primary header-nav">구매 후기</a>
         <button onclick="logout()" class="btn btn-danger header-nav">로그아웃</button>
     </div>
+
+    <script>
+        // 세션 스토리지에서 토큰을 가져옴
+        var adminToken = sessionStorage.getItem("adminToken");
+        
+        // 토큰이 없으면 로그인 페이지로 리다이렉트
+        if (!adminToken) {
+            window.location.href = "admin?command=admin_login_form";
+        }
+    </script>
 </body>
 </html>
