+++ src/main/webapp/views/includes/main_slide.jsp
@@ -1,15 +1,9 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
 	pageEncoding="UTF-8"%>
-<div id="slider">
-	<img src="/resources/img/slide_image/membership_img.jpg" alt="회원가입혜택"
-		id="memberShip"> <img
-		src="/resources/img/slide_image/mobile_img.jpg" alt="모바일앱쿠폰이벤트"
-		id="mobileAward"> <img
-		src="/resources/img/slide_image/delivery_img.jpg" alt="배송당일출발"
-		id="delivery">
-	<ul>
-		<li><a href="#">1</a></li>
-		<li><a href="#">2</a></li>
-		<li><a href="#">3</a></li>
-	</ul>
+<div class="container">
+    <div class="fade-in-out-slider">
+        <img id="mobileAward" src="/resources/img/slide_image/mobile_img.jpg" alt="모바일앱쿠폰이벤트" class="active">
+        <img id="delivery" src="/resources/img/slide_image/delivery_img.jpg" alt="배송당일출발">
+        <img id="membership" src="/resources/img/slide_image/membership_img.jpg" alt="배송당일출발">
+    </div>
 </div>
