+++ src/main/webapp/views/includes/top_banner.jsp
@@ -1,11 +1,5 @@
-
 	<!-- /* top banner */ -->
-	
-	<div class="topBack">
 		<div id="topBanner">
-			<a> <img src="/resources/img/top_banner/top_banner.jpg">
-			</a>
+			<a><img src="/resources/img/top_banner/top_banner.jpg"></a>
 			<button id="topClose"></button>
 		</div>
-	</div>
-
