+++ src/main/webapp/views/index.jsp
@@ -10,98 +10,36 @@
 <%@include file="includes/cart.jsp"%>
 <%@include file="includes/footer.jsp"%>
 
+
 <script>
-        /* slider */
-        /* 메인 슬라이더 페이드 인 페이드 아웃 기능 
-        v0.01 : 240107 setTime()메서드에 시간을 입력하며 하드 코딩함
-        v0.02 : 240110 로직을 조금 만들어 함수를 활용하지만 개체수 증가시 코드도 함께 증가
-        v0.03 : 240115 개체수 증가시 코드 증가 문제 해결 -> 이미지 추가 시 아이디만 배열에 추가
-        v0.04 : 240119 웹페이지 포커스 대기간 쌓이는 인터벌로 인한 페이드 인, 아웃 이상 현상 개선/visiblechange이벤트 활용
-        */
-        
-        function fadeInManager() {
+    document.addEventListener("DOMContentLoaded", function() {
+        let images = document.querySelectorAll('.fade-in-out-slider img');
+        let index = 0;
 
-            let opacityValueFadeIn = 0;
-            function FadeInUnit(image) {
-                this.image = image;
-                this.fadeIn = function () {
-                    document.getElementById(image).style.opacity = opacityValueFadeIn;
-                    opacityValueFadeIn += 0.1;
-                    if (opacityValueFadeIn >= 1.5) {
-                        clearInterval(myFadeIn);
-                        opacityValueFadeIn = 0;
-                    }
-                }
-                let myFadeIn = setInterval(this.fadeIn, 100);
-            }
+        function fadeInOut() {
+            
+            images[index].classList.remove('active');
 
-            let imageIdIndexFadeIn = 0;
-            function repeatFadeIn() {
-                const imageId = ["mobileAward", "delivery"];
-                if (imageIdIndexFadeIn == imageId.length) {
-                    stopFadeInInterval();
-                    return
-                }
-                let createFadeIn = new FadeInUnit(imageId[imageIdIndexFadeIn]);
-                imageIdIndexFadeIn++;
-            }
+            index = (index + 1) % images.length;
 
-            let fadeInInterval = setInterval(repeatFadeIn, 2500);
-
-            function stopFadeInInterval() {
-                clearInterval(fadeInInterval);
-            }
+            images[index].classList.add('active');
         }
 
-        function fadeOutManager() {
+        images[index].classList.add('active');
 
-            let opacityValueFadeOut = 1;
-            function FadeOutUnit(image) {
-                this.image = image;
-                this.fadeOut = function () {
-                    document.getElementById(image).style.opacity = opacityValueFadeOut;
-                    opacityValueFadeOut -= 0.1;
-                    if (opacityValueFadeOut <= -0.7) {
-                        clearInterval(myFadeOut);
-                        opacityValueFadeOut = 1;
-                    }
-                }
-                let myFadeOut = setInterval(this.fadeOut, 100);
-            }
+        // Set interval to call fadeInOut function every 5 seconds
+        let interval = setInterval(fadeInOut, 2000);
 
-            let imageIdIndexFadeOut = 1;
-            function repeatFadeOut() {
-                const imageId = ["mobileAward", "delivery"];
-                if (imageIdIndexFadeOut < 0) {
-                    stopFadeOutInterval();
-                    return
-                }
-                let createFadeOut = new FadeOutUnit(imageId[imageIdIndexFadeOut]);
-                imageIdIndexFadeOut--;
-            }
+        // Clear interval when mouse hovers over the slider
+        document.querySelector('.fade-in-out-slider').addEventListener('mouseenter', function() {
+            clearInterval(interval);
+        });
 
-            let fadeOutInterval = setInterval(repeatFadeOut, 2500);
-
-            function stopFadeOutInterval() {
-                clearInterval(fadeOutInterval);
-            }
-        }
-
-        let sliderTimer;
-
-        function repeat() {
-            setTimeout(fadeOutManager, 0);
-            setTimeout(fadeInManager, 6000);
-        }
-
-        function sliderStart() {
-            setTimeout(repeat, 0);
-            sliderTimer = setInterval(repeat, 12000);
-        }
-
-        function sliderStop() {
-            clearInterval(sliderTimer);
-        }
-
-        sliderStart();
+        document.querySelector('.fade-in-out-slider').addEventListener('mouseleave', function() {
+            interval = setInterval(fadeInOut, 2000);
+        });
+    });
 </script>
+
+</body>
+</html>
